Renamed api model manager.
