+++ tslint.json
@@ -15,6 +15,10 @@
         "eofline": true,
         "indent": [true, "spaces", 2],
         "interface-name": [true, "never-prefix"],
+        "max-line-length": [true, {
+          "limit": 120,
+          "ignore-pattern": "^import .+ from"
+        }],
         "no-empty-interface": false,
         "no-var-keyword": false,
         "no-var-requires": false,
